requirements.txt
django>=1.8,<1.9
django-debug-toolbar==1.5
gunicorn==19.4.5
whitenoise==3.0
Pillow=4.0.0
django-sendgrid=4.0.4
